# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "E-commerce MVP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List Products",
        "operationId": "list_products_api_products_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search on title/description",
              "title": "Q"
            },
            "description": "Free-text search on title/description"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "min_price_cents",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Price Cents"
            }
          },
          {
            "name": "max_price_cents",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Price Cents"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create Product",
        "operationId": "create_product_api_products_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{product_id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get Product",
        "operationId": "get_product_api_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "products"
        ],
        "summary": "Update Product",
        "operationId": "update_product_api_products__product_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{product_id}/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Create Review",
        "operationId": "create_review_api_products__product_id__reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_api_products__product_id__reviews_get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewOut"
                  },
                  "title": "Response List Reviews Api Products  Product Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderOut"
                  },
                  "type": "array",
                  "title": "Response List Orders Api Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/seller/products": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "My Products",
        "operationId": "my_products_api_seller_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "type": "array",
                  "title": "Response My Products Api Seller Products Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/seller/orders": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "Received Orders",
        "operationId": "received_orders_api_seller_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/seller/dashboard": {
      "get": {
        "tags": [
          "seller"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_seller_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerDashboard"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "title_snapshot": {
            "type": "string",
            "title": "Title Snapshot"
          },
          "unit_price_cents": {
            "type": "integer",
            "title": "Unit Price Cents"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "title_snapshot",
          "unit_price_cents",
          "quantity"
        ],
        "title": "OrderItemOut"
      },
      "OrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "buyer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Buyer Id"
          },
          "total_cents": {
            "type": "integer",
            "title": "Total Cents"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "payment_status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "shipment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShipmentOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "buyer_id",
          "total_cents",
          "status",
          "payment_status",
          "created_at",
          "items"
        ],
        "title": "OrderOut"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "PAID",
          "SHIPPED",
          "DELIVERED",
          "CANCELLED"
        ],
        "title": "OrderStatus"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "UNPAID",
          "PAID",
          "REFUNDED"
        ],
        "title": "PaymentStatus"
      },
      "ProductCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "price_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "BRL"
          },
          "stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "price_cents",
          "stock",
          "category"
        ],
        "title": "ProductCreate"
      },
      "ProductListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ProductListOut"
      },
      "ProductOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seller Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "price_cents": {
            "type": "integer",
            "title": "Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "stock": {
            "type": "integer",
            "title": "Stock"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "seller": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SellerInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "seller_id",
          "title",
          "description",
          "price_cents",
          "currency",
          "stock",
          "category",
          "images",
          "status",
          "created_at"
        ],
        "title": "ProductOut"
      },
      "ProductStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "PAUSED"
        ],
        "title": "ProductStatus"
      },
      "ProductUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Cents"
          },
          "stock": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
...
      "ReviewCreate": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "title": "Comment",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "ReviewCreate"
      },
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "buyer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Buyer Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "product_id",
          "buyer_id",
          "rating",
          "comment",
          "created_at"
        ],
        "title": "ReviewOut"
      },
      "SellerDashboard": {
        "properties": {
          "store_name": {
            "type": "string",
            "title": "Store Name"
          },
          "reputation_score": {
            "type": "number",
            "title": "Reputation Score"
          },
          "active_products": {
            "type": "integer",
            "title": "Active Products"
          },
          "total_orders": {
            "type": "integer",
            "title": "Total Orders"
          },
          "revenue_cents": {
            "type": "integer",
            "title": "Revenue Cents"
          }
        },
        "type": "object",
        "required": [
          "store_name",
          "reputation_score",
          "active_products",
          "total_orders",
          "revenue_cents"
        ],
        "title": "SellerDashboard"
      },
      "SellerInfo": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "store_name": {
            "type": "string",
            "title": "Store Name"
          },
          "reputation_score": {
            "type": "number",
            "title": "Reputation Score"
          },
          "sales_count": {
            "type": "integer",
            "title": "Sales Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "store_name",
          "reputation_score",
          "sales_count"
        ],
        "title": "SellerInfo"
      },
      "ShipmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "status": {
            "$ref": "#/components/schemas/ShipmentStatus"
          },
          "tracking_code": {
            "type": "string",
            "title": "Tracking Code"
          },
          "estimated_delivery": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Delivery"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "status",
          "tracking_code",
          "estimated_delivery",
          "updated_at"
        ],
        "title": "ShipmentOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}